@@ -1,5 +1,4 @@
-import mongodb from "mongodb"
-const ObjectId = mongodb.ObjectID
+import { ObjectId } from "mongodb"
 
 let reviews
 
@@ -9,7 +8,7 @@ export default class ReviewsDAO {
       return
     }
     try {
-      reviews = await conn.db(process.env.RESTREVIEWS_NS).collection("reviews")
+      reviews = await conn.db(process.env.RESTREVIEWS_NS || "restaurant_reviews").collection("reviews")
     } catch (e) {
       console.error(`Unable to establish collection handles in userDAO: ${e}`)
     }
@@ -21,7 +20,7 @@ export default class ReviewsDAO {
           user_id: user._id,
           date: date,
           text: review,
-          restaurant_id: ObjectId(restaurantId), }
+          restaurant_id: new ObjectId(restaurantId), }
 
       return await reviews.insertOne(reviewDoc)
     } catch (e) {
@@ -33,7 +32,7 @@ export default class ReviewsDAO {
   static async updateReview(reviewId, userId, text, date) {
     try {
       const updateResponse = await reviews.updateOne(
-        { user_id: userId, _id: ObjectId(reviewId)},
+        { user_id: userId, _id: new ObjectId(reviewId)},
         { $set: { text: text, date: date  } },
       )
 
@@ -48,7 +47,7 @@ export default class ReviewsDAO {
 
     try {
       const deleteResponse = await reviews.deleteOne({
-        _id: ObjectId(reviewId),
+        _id: new ObjectId(reviewId),
         user_id: userId,
       })
 
