@@ -1,15 +1,17 @@
 import React, { useState } from "react";
 import RestaurantDataService from "../services/restaurant";
-import { Link } from "react-router-dom";
+import { Link, useParams, useLocation } from "react-router-dom";
 
 const AddReview = props => {
+  const { id } = useParams();
+  const location = useLocation();
   let initialReviewState = ""
 
   let editing = false;
 
-  if (props.location.state && props.location.state.currentReview) {
+  if (location.state && location.state.currentReview) {
     editing = true;
-    initialReviewState = props.location.state.currentReview.text
+    initialReviewState = location.state.currentReview.text
   }
 
   const [review, setReview] = useState(initialReviewState);
@@ -24,11 +26,11 @@ const AddReview = props => {
       text: review,
       name: props.user.name,
       user_id: props.user.id,
-      restaurant_id: props.match.params.id
+      restaurant_id: id
     };
 
     if (editing) {
-      data.review_id = props.location.state.currentReview._id
+      data.review_id = location.state.currentReview._id
       RestaurantDataService.updateReview(data)
         .then(response => {
           setSubmitted(true);
@@ -57,7 +59,7 @@ const AddReview = props => {
         {submitted ? (
           <div>
             <h4>You submitted successfully!</h4>
-            <Link to={"/restaurants/" + props.match.params.id} className="btn btn-success">
+            <Link to={"/restaurants/" + id} className="btn btn-success">
               Back to Restaurant
             </Link>
           </div>
