@@ -1,6 +1,8 @@
 import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
 const Login = props => {
+  const navigate = useNavigate();
 
   const initialUserState = {
     name: "",
@@ -16,7 +18,7 @@ const Login = props => {
 
   const login = () => {
     props.login(user)
-    props.history.push('/');
+    navigate('/');
   }
 
   return (
